25_06_29_12_12_00 pybear-dask separation

+++ tests/model_selection/GSTCV/GSTCVDask/fit/accuracy_50sec_test.py
@@ -110,7 +110,7 @@ class TestFitAccuracy:
         TestCls = GSTCVDask(
             estimator=special_sk_est_log,
             param_grid=_param_grid,
-            thresholds=np.float64(0.5),
+            thresholds=0.5,
             cv=_cv_iter,
             error_score=standard_error_score,
             refit=False,
